feat: 문의 생성 api 추가 - 파일 업로드 추가

+++ src/main/java/one/dfy/bily/api/admin/service/SpaceService.java
@@ -260,7 +260,7 @@ public class SpaceService {
 
     }
 
-    public Space findById(Long contentId) {
+    public Space findById(Integer contentId) {
         if (contentId == null) {
             return null;
         }
